requirements.txt
playwright
beautifulsoup4
pandas
requests
python-dotenv
aiofiles
